+++ CHANGELOG.md
@@ -22,6 +22,8 @@ We have encountered transient lambda service errors in our integration testing. 
   - Adds new BackendApi endpoint `distributionMetrics` that returns a summary of successful s3 accesses as well as a summary of distribution errors -- including s3 access errors, 4XX and 5XX errors.
 - **CUMULUS-1273**
   - Added lambda function EmsProductMetadataReport to generate EMS Product Metadata report
+- **CUMULUS-824**
+  - Added new Collection parameter `reportToEms` to configure whether the collection is reported to EMS
 - **CUMULUS-1241**
   - Added information about queues with maximum execution limits defined to default workflow templates (`meta.queueExecutionLimits`)
 - **CUMULUS-1311**